hotfix: rename thanosql to thanosql_magic for
