+++ thanosql_magic/__init__.py
