+++ thanosql_magic/exception.py
